XMLHttpRequest/responsexml-basic.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: responseXML basic test</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function() {
        var client = new XMLHttpRequest()
        client.open("GET", "resources/well-formed.xml", false)
        client.send(null)
        assert_equals(client.responseXML.documentElement.localName, "html")
        assert_equals(client.responseXML.documentElement.childNodes.length, 5)
        assert_equals(client.responseXML.getElementById("n1").localName, client.responseXML.documentElement.childNodes[1].localName)
        assert_equals(client.responseXML.getElementById("n2"), null)
        assert_equals(client.responseXML.getElementsByTagName("p")[1].namespaceURI, "namespacesarejuststrings")
        alert(client.responseXML.lastModified)
        assert_equals(client.responseXML.domain, "")
        client.responseXML.cookie = "thisshouldbeignored"
        assert_equals(client.responseXML.cookie, "")
      })
    </script>
  </body>
</html>
